Feed raw input data to libreadline
In the past we were forced to convert a Key event as we received it
through the UI back into a byte sequence that libreadline could
understand. For common printable ASCII characters that is typically not
a huge problem, but for escape sequences which can easily be terminal
specific, that is a very big deal. Because we have no interest in
translating every possible key back, we ended up just supporting a very
basic set: backspace, delete, and escape.
Now that the gui crate supports arbitrary event types, we are using an
event type that contains the raw input data as handled by termion. With
this patch we use this raw input data and feed it to libreadline. Doing
so we can remove the cumbersome back translation logic.