Simplify common creation of Event objects
The majority of the UI tests use the Event type in one form or another.
A common task is to create an event representing a particular key to
send it to the UI.
Given how many tests perform such an action, it makes sense to simplify
the creation of such Event objects. Hence, with this change we introduce
additional From conversion for this type that shorten the construction
from characters, bytes, and keys.